move some classes and definitions
